Create a trait for public identifiers
