+++ src/DataService/Player/Player.php
@@ -6,6 +6,7 @@ use Lullabot\Mpx\DataService\Annotation\DataService;
 use Lullabot\Mpx\DataService\AdPolicyDataTrait;
 use Lullabot\Mpx\DataService\ObjectBase;
 use Lullabot\Mpx\DataService\PublicIdentifierInterface;
+use Lullabot\Mpx\DataService\PublicIdentifierTrait;
 use Psr\Http\Message\UriInterface;
 
 /**
@@ -19,6 +20,7 @@ use Psr\Http\Message\UriInterface;
 class Player extends ObjectBase implements PublicIdentifierInterface
 {
     use AdPolicyDataTrait;
+    use PublicIdentifierTrait;
 
     /**
      * The administrative workflow tags for this object.
@@ -397,13 +399,6 @@ class Player extends ObjectBase implements PublicIdentifierInterface
      * @var string
      */
     protected $pdk;
-
-    /**
-     * The public identifier for this player when requested through the Player Service.
-     *
-     * @var string
-     */
-    protected $pid;
 
     /**
      * Indicates if the player should automatically play the next release when one finishes.
@@ -1686,26 +1681,6 @@ class Player extends ObjectBase implements PublicIdentifierInterface
     public function setPdk($pdk)
     {
         $this->pdk = $pdk;
-    }
-
-    /**
-     * Returns the public identifier for this player when requested through the Player Service.
-     *
-     * @return string
-     */
-    public function getPid(): string
-    {
-        return $this->pid;
-    }
-
-    /**
-     * Set the public identifier for this player when requested through the Player Service.
-     *
-     * @param string
-     */
-    public function setPid($pid)
-    {
-        $this->pid = $pid;
     }
 
     /**
